% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-usa-federal.R
\name{cal_usa_federal}
\alias{cal_usa_federal}
\title{Calendar - USA Federal}
\usage{
cal_usa_federal(since = "1900-01-01", until = "2100-01-01", observed = TRUE)
}
\arguments{
\item{since}{\verb{[Date(1)]}

The lower bound on where to begin looking for the holiday.}

\item{until}{\verb{[Date(1)]}

The upper bound on where to end looking for the holiday.}

\item{observed}{\verb{[logical(1)]}

If the holiday falls on a weekend, should the observed date be returned?
The observed date is the nearest weekday. If \code{FALSE}, the exact holiday
date is returned.}
}
\value{
An rcalendar.
}
\description{
\code{cal_usa_federal()} is a federal calendar for the United States.
}
\section{Holidays}{

\itemize{
\item Martin Luther King Jr. Day
\item Christmas
\item New Year's Day
}
}

\examples{
library(almanac)

cal <- cal_usa_federal()

alma_search("2000-01-01", "2001-01-01", cal)
}
