% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hldy-martin-luther-king-jr-day.R
\name{hldy_martin_luther_king_jr_day}
\alias{hldy_martin_luther_king_jr_day}
\title{Christmas}
\usage{
hldy_martin_luther_king_jr_day(
  since = "1900-01-01",
  until = "2100-01-01",
  adjust_on = NULL,
  adjustment = NULL
)
}
\arguments{
\item{since}{\verb{[Date(1)]}

The lower bound on where to begin looking for the holiday.}

\item{until}{\verb{[Date(1)]}

The upper bound on where to end looking for the holiday.}

\item{adjust_on}{\verb{[NULL / rschedule]}

If \code{NULL}, no adjustment is made to the holiday.

Otherwise, an rschedule that defines when an adjustment to the holiday
should be made. For example, set to an rschedule for "on weekends", and
supply an \code{adjustment} of \code{adj_nearest()} to roll the holiday to the
nearest weekday.}

\item{adjustment}{\verb{[NULL / function]}

If \code{NULL}, no adjustment is made to the holiday.

Otherwise, an adjustment function to apply to problematic dates. Typically
one of the pre-existing adjustment functions, like
\code{\link[almanac:adj_nearest]{almanac::adj_nearest()}}.}
}
\value{
A rholiday.
}
\description{
\code{hldy_martin_luther_king_jr_day()} is a holiday for Martin Luther King
Jr. Day, the third Monday in January.
}
\examples{
alma_search("1990-01-01", "1995-01-01", hldy_martin_luther_king_jr_day())
}
