% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacation-package.R
\docType{package}
\name{vacation-package}
\alias{vacation}
\alias{vacation-package}
\title{vacation: Holiday and Calendar Extensions for 'almanac'}
\description{
Provides additional holiday and calendar utilities
    extending the recurrence rules tools provided by the 'almanac'
    package. While 'almanac' provides the core tools for working with
    recurring events, it is a little challenging to use it to construct a
    full business calendar. The goal of 'vacation' is to provide an
    extensible system for adding custom holidays into a larger business
    calendar. 'vacation' comes with a number of common holidays built-in.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/vacation}
  \item Report bugs at \url{https://github.com/DavisVaughan/vacation/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

}
\keyword{internal}
